It's done!

+++ src/main/java/lang/print/gaps/task6/FloatTypeCasting.java
@@ -3,6 +3,6 @@ package lang.print.gaps.task6;
 public class FloatTypeCasting {
     public void roundNumber(float numberToBeRounded) {
         numberToBeRounded = Math.round(numberToBeRounded);
-        System.out.println(numberToBeRounded);
+        System.out.println((int)numberToBeRounded);
     }
 }
